Added xion crate instead of local xion (#525)
