+++ framework/Cargo.toml
@@ -86,7 +86,10 @@ abstract-ica = { version = "0.24.1", path = "packages/abstract-ica" }
 abstract-sdk = { version = "0.24.1", path = "packages/abstract-sdk" }
 abstract-testing = { version = "0.24.1", path = "packages/abstract-testing" }
 abstract-std = { version = "0.24.1", path = "packages/abstract-std" }
-abstract-xion = { version = "0.24.1", path = "packages/abstract-xion" }
+abstract-xion = { package = "account", version = "0.1.0", git = "https://github.com/abstractsdk/xion-contracts", features = [
+  "library",
+], branch = "fix-for-abstract" }
+
 
 # These should remain fixed and don't need to be re-published (unless changes are made)
 abstract-macros = { version = "0.24.1", path = "packages/abstract-macros" }
