+++ framework/contracts/account/src/contract.rs
@@ -9,7 +9,10 @@ use abstract_sdk::{
 };
 use abstract_std::{
     account::{
-        state::{AccountInfo, WhitelistedModules, INFO, SUSPENSION_STATUS, WHITELISTED_MODULES},
+        state::{
+            AccountInfo, WhitelistedModules, AUTH_ADMIN, INFO, SUSPENSION_STATUS,
+            WHITELISTED_MODULES,
+        },
         UpdateSubAccountAction,
     },
     module_factory::SimulateInstallModulesResponse,
@@ -163,7 +166,7 @@ pub fn instantiate(
                 let Some(mut add_auth) = authenticator else {
                     return Err(AccountError::AbsAccNoAuth {});
                 };
-                abstract_xion::auth::execute::add_auth_method(deps.branch(), &env, &mut add_auth)?;
+                abstract_xion::execute::add_auth_method(deps.branch(), &env, &mut add_auth)?;
 
                 response = response.add_event(
                     cosmwasm_std::Event::new("create_abstract_account").add_attributes(vec![
@@ -258,16 +261,17 @@ pub fn instantiate(
 
 #[cfg_attr(feature = "export", cosmwasm_std::entry_point)]
 pub fn execute(mut deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> AccountResult {
-    match msg {
+    let response = match msg {
         ExecuteMsg::UpdateStatus {
             is_suspended: suspension_status,
-        } => update_account_status(deps, info, suspension_status),
+        } => update_account_status(deps.branch(), info, suspension_status),
         msg => {
             // Block actions if account is suspended
             let is_suspended = SUSPENSION_STATUS.load(deps.storage)?;
             if is_suspended {
                 return Err(AccountError::AccountSuspended {});
             }
+            let mut deps = deps.branch();
 
             match msg {
                 // ## Execution ##
@@ -367,7 +371,9 @@ pub fn execute(mut deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) 
                 } => send_funds_with_actions(deps, info, env, host_chain, amount, actions),
             }
         }
-    }
+    }?;
+    AUTH_ADMIN.remove(deps.storage);
+    Ok(response)
 }
 
 #[cfg_attr(feature = "export", cosmwasm_std::entry_point)]
@@ -405,13 +411,18 @@ pub fn query(deps: Deps, env: Env, msg: QueryMsg) -> StdResult<Binary> {
         #[cfg_attr(not(feature = "xion"), allow(unused_variables))]
         QueryMsg::AuthenticatorByID { id } => {
             #[cfg(feature = "xion")]
-            return abstract_xion::queries::authenticator_by_id(deps.storage, id);
+            return cosmwasm_std::to_json_binary(&abstract_xion::query::authenticator_by_id(
+                deps.storage,
+                id,
+            )?);
             #[cfg(not(feature = "xion"))]
             Ok(Binary::default())
         }
         QueryMsg::AuthenticatorIDs {} => {
             #[cfg(feature = "xion")]
-            return abstract_xion::queries::authenticator_ids(deps.storage);
+            return cosmwasm_std::to_json_binary(&abstract_xion::query::authenticator_ids(
+                deps.storage,
+            )?);
             #[cfg(not(feature = "xion"))]
             Ok(Binary::default())
         }
@@ -424,7 +435,7 @@ pub fn sudo(deps: DepsMut, env: Env, msg: SudoMsg) -> AccountResult {
         SudoMsg::IBCLifecycleComplete(msg) => ics20_hook_callback(deps, env, msg),
 
         #[cfg(feature = "xion")]
-        SudoMsg::Xion(msg) => abstract_xion::sudo::sudo(deps, env, msg).map_err(Into::into),
+        SudoMsg::Xion(msg) => abstract_xion::contract::sudo(deps, env, msg).map_err(Into::into),
     }
 }
