+++ framework/contracts/account/src/error.rs
@@ -129,7 +129,7 @@ pub enum AccountError {
 
     #[cfg(feature = "xion")]
     #[error(transparent)]
-    AbstractXion(#[from] abstract_xion::error::AbstractXionError),
+    AbstractXion(#[from] abstract_xion::error::ContractError),
 
     #[error("Chain {0} not registered in the IBC Client")]
     ChainNotRegistered(TruncatedChainId),
