+++ framework/contracts/account/src/execution.rs
@@ -145,7 +145,7 @@ pub fn add_auth_method(
 ) -> AccountResult {
     #[cfg(feature = "xion")]
     {
-        abstract_xion::auth::execute::add_auth_method(_deps, &_env, &mut _auth).map_err(Into::into)
+        abstract_xion::execute::add_auth_method(_deps, &_env, &mut _auth).map_err(Into::into)
     }
     #[cfg(not(feature = "xion"))]
     {
@@ -156,7 +156,7 @@ pub fn add_auth_method(
 pub fn remove_auth_method(_deps: DepsMut, _env: Env, _id: u8) -> AccountResult {
     #[cfg(feature = "xion")]
     {
-        abstract_xion::auth::execute::remove_auth_method(_deps, _env, _id).map_err(Into::into)
+        abstract_xion::execute::remove_auth_method(_deps, _env, _id).map_err(Into::into)
     }
     #[cfg(not(feature = "xion"))]
     {
