--- framework/packages/abstract-xion/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "abstract-xion"
-version = "0.24.2"
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-documentation.workspace = true
-repository.workspace = true
-license.workspace = true
-keywords.workspace = true
-description = "Abstract account implementation on xion chain"
-
-[dependencies]
-abstract-std = { workspace = true }
-cosmwasm-schema = { workspace = true }
-cosmwasm-std = { workspace = true, features = ["cosmwasm_2_0"] }
-cw-storage-plus = { workspace = true }
-thiserror = { workspace = true }
-
-tiny-keccak = { version = "2", default-features = false, features = ["keccak"] }
-hex = { version = "0.4", default-features = false }
-sha2 = { version = "0.10.8", default-features = false, features = ["oid"] }
-bech32 = { version = "0.11.0" }
-base64 = { version = "0.22.1", default-features = false }
-rsa = { version = "0.9.6", default-features = false, features = ["std"] }
-getrandom = { version = "0.2.15", default-features = false, features = [
-    "custom",
-] }
-p256 = { version = "0.13.2", features = ["ecdsa-core", "arithmetic", "serde"] }
-url = { version = "2.5.2", default-features = false }
-ripemd = { version = "0.1.3", default-features = false }
-prost = { version = "0.13.2" }
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-workspace-hack = { version = "0.1", path = "../../workspace-hack" }
-
-[lints]
-workspace = true
-
-[dev-dependencies]
-coverage-helper = { workspace = true }
